# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "soca web api v1.0",
        "description": "soca web API Description",
        "version": "1.0"
    },
    "paths": {
        "/api/Deps/{pg}": {
            "get": {
                "tags": [
                    "部門"
                ],
                "summary": "Get Deps data",
                "operationId": "07bc341cbd1f79b0a7d73356f436795a",
                "parameters": [
                    {
                        "name": "pg",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of 部門"
                    }
                }
            }
        },
        "/api/history/{pg}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get a list of history",
                "operationId": "0682611335fd6cb088585fedb3c6cd68",
                "parameters": [
                    {
                        "name": "pg",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of history"
                    }
                }
            }
        },
...
        "/api/history/": {
            "post": {
                "tags": [
                    "History"
                ],
                "summary": "Create History data",
                "operationId": "f3d31f731d4ada9121c20ce7b79fb409",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "R_NO": {
                                        "description": "Reader number",
                                        "type": "integer"
                                    },
                                    "R_MODEL": {
                                        "description": "Reader model",
                                        "type": "integer"
                                    },
                                    "R_NAME": {
                                        "description": "Reader name",
                                        "type": "string"
                                    },
                                    "R_INTERFACE": {
                                        "description": "Reader interface",
                                        "type": "string"
                                    },
                                    "R_TIMEOUT": {
                                        "description": "Reader timeout",
                                        "type": "integer"
                                    },
                                    "R_MAP_IDX": {
                                        "description": "Reader map index",
                                        "type": "integer"
                                    },
                                    "R_READ_POINTER3": {
                                        "description": "Reader read pointer 3",
                                        "type": "integer"
                                    },
                                    "R_READ_POINTER4": {
                                        "description": "Reader read pointer 4",
                                        "type": "integer"
                                    },
                                    "POLLING": {
                                        "description": "Polling value",
                                        "type": "integer"
                                    },
                                    "CREATETIME": {
                                        "description": "Creation time",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Delete historys Id"
                    },
                    "500": {
                        "description": "Failed to create historys"
                    }
                }
            }
        },
...
    },
...
}